Refactor the code.
